refactor: render the whole animation of spinning logo in threejs
