+++ src/components/Home/SpinningLogo.tsx
@@ -1,40 +1,151 @@
-import Image from "next/image";
+import * as THREE from "three";
+import { useEffect, useRef } from "react";
 
 export default function SpinningLogo() {
+  const refContainer = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const scene = new THREE.Scene();
+    const radius = 1.6;
+
+    const camera = new THREE.PerspectiveCamera(
+      75,
+      window.innerWidth / window.innerHeight,
+      0.1,
+      1000
+    );
+    camera.position.set(0, 0, 5);
+
+    const renderer = new THREE.WebGLRenderer({ alpha: true, antialias: true });
+    renderer.setSize(window.innerWidth, window.innerHeight);
+    renderer.setPixelRatio(window.devicePixelRatio);
+    renderer.shadowMap.enabled = true;
+
+    if (refContainer.current) {
+      refContainer.current.appendChild(renderer.domElement);
+    }
+
+    const setRendererSize = () => {
+      if (refContainer.current) {
+        renderer.setSize(
+          refContainer.current.clientWidth,
+          refContainer.current.clientHeight
+        );
+        camera.aspect =
+          refContainer.current.clientWidth / refContainer.current.clientHeight;
+        camera.updateProjectionMatrix();
+      }
+    };
+    window.addEventListener("resize", setRendererSize);
+
+    const ambientLight = new THREE.AmbientLight(0xc14627, 0.5);
+    scene.add(ambientLight);
+
+    const pointLight = new THREE.PointLight(0xc14627, 2, 20);
+    pointLight.position.set(0, 0, 5);
+    scene.add(pointLight);
+
+    const circleCurve = new THREE.CatmullRomCurve3(
+      new Array(100).fill(0).map((_, i) => {
+        const angle = (i / 100) * Math.PI * 2;
+        return new THREE.Vector3(
+          Math.cos(angle) * radius,
+          Math.sin(angle) * radius,
+          0
+        );
+      }),
+      true
+    );
+
+    const tubeGeometry = new THREE.TubeGeometry(
+      circleCurve,
+      100,
+      0.07,
+      16,
+      true
+    );
+
+    const tubeMaterial = new THREE.MeshStandardMaterial({
+      color: 0xc14627,
+      transparent: true,
+      opacity: 0.5,
+      emissive: 0xc14627,
+      emissiveIntensity: 1,
+      metalness: 0.8,
+      roughness: 0.3,
+    });
+
+    const tube = new THREE.Mesh(tubeGeometry, tubeMaterial);
+    scene.add(tube);
+
+    const textureLoader = new THREE.TextureLoader();
+    const baseTexture = textureLoader.load("/hologram/hologram_nolight.png");
+    const lightTextures = [
+      textureLoader.load("/hologram/hologram_light1.png"),
+      textureLoader.load("/hologram/hologram_light2.png"),
+      textureLoader.load("/hologram/hologram_light3.png"),
+    ];
+
+    const frontGeometry = new THREE.PlaneGeometry(radius * 2, radius * 2);
+    const logoMaterials = [
+      new THREE.MeshBasicMaterial({
+        map: baseTexture,
+        transparent: true,
+        opacity: 1,
+      }),
+      ...lightTextures.map(
+        (texture) =>
+          new THREE.MeshBasicMaterial({
+            map: texture,
+            transparent: true,
+            opacity: 0,
+          })
+      ),
+    ];
+
+    const frontMeshes = logoMaterials.map(
+      (mat) => new THREE.Mesh(frontGeometry, mat)
+    );
+
+    frontMeshes.forEach((mesh) => {
+      mesh.position.set(0, 0, 0.1);
+      scene.add(mesh);
+    });
+
+    const backMeshes = frontMeshes.map((mesh) => {
+      const backMesh = mesh.clone();
+      backMesh.rotation.y = Math.PI;
+      backMesh.position.set(0, 0, 0.1);
+      scene.add(backMesh);
+      return backMesh;
+    });
+
+    const allPlaneMeshes = [...frontMeshes, ...backMeshes];
+
+    function flickerLights() {
+      logoMaterials.slice(1).forEach((material) => {
+        material.opacity = Math.random() > 0.5 ? 1 : 0;
+      });
+
+      setTimeout(flickerLights, Math.random() * 500);
+    }
+
+    flickerLights();
+
+    const animate = function () {
+      requestAnimationFrame(animate);
+      tube.rotation.y += 0.01;
+      allPlaneMeshes.forEach((mesh) => (mesh.rotation.y += 0.01));
+      renderer.render(scene, camera);
+    };
+
+    animate();
+  }, []);
+
   return (
-    <div className="absolute top-[calc(50vh-12rem)] left-[calc(50vw-19.2rem)] -translate-x-1/2 -translate-y-1/2 h-96 w-[38.4rem] z-10 transform-style-3d hidden md:block animate-spin-pulse">
-      <Image
-        src="/hologram/hologram_nolight.png"
-        alt="Hologram Base"
-        width={500}
-        height={500}
-        className="absolute top-0 left-0 h-full w-auto object-contain"
-        onError={(e) => console.error("Failed to load image:", e)}
-      />
-      <Image
-        src="/hologram/hologram_light1.png"
-        alt="Hologram Light 1"
-        width={500}
-        height={500}
-        className="absolute top-0 left-0 h-full w-auto object-contain animate-flicker-1"
-        onError={(e) => console.error("Failed to load image:", e)}
-      />
-      <Image
-        src="/hologram/hologram_light2.png"
-        alt="Hologram Light 2"
-        width={500}
-        height={500}
-        className="absolute top-0 left-0 h-full w-auto object-contain animate-flicker-2"
-        onError={(e) => console.error("Failed to load image:", e)}
-      />
-      <Image
-        src="/hologram/hologram_light3.png"
-        alt="Hologram Light 3"
-        width={500}
-        height={500}
-        className="absolute top-0 left-0 h-full w-auto object-contain animate-flicker-3"
-        onError={(e) => console.error("Failed to load image:", e)}
-      />
-    </div>
+    <div
+      ref={refContainer}
+      className="fixed top-0 left-0 w-full h-full overflow-hidden hidden md:block -z-10"
+    ></div>
   );
 }
